pyproject.toml
[project]
name = "agents"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "boto3>=1.42.59",
    "fastapi>=0.135.1",
    "litellm>=1.80.0",
    "motor>=3.7.1",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
    "qdrant-client>=1.17.0",
    "requests>=2.32.5",
    "strands-agents[openai]>=1.28.0",
    "streamlit>=1.54.0",
    "uvicorn[standard]>=0.41.0",
]
